requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pymysql
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
